Create Hashable module and integrate into classes
