+++ lib/game_teams.rb
@@ -1,8 +1,10 @@
 require 'csv'
 require_relative 'modules/groupable'
+require_relative 'modules/hashable'
 
 class GameTeams 
   include Groupable
+  include Hashable
   attr_reader :game_teams_path, :game_path, :team_path
 
   
@@ -24,34 +26,6 @@ class GameTeams
     end
     average_goals_per_team
   end
-
-  def visitor_scores_hash
-		games_grouped_by_away_team = @game_teams_path.group_by {|row| row[:team_id]}
-		average_away_goals_per_team = {}
-
-		games_grouped_by_away_team.each do |team, games|
-			average_away_goals_per_team[team] = 0
-				games.each do |game|
-				average_away_goals_per_team[team] += game[:goals].to_i if game[:hoa] == "away"
-			end
-			average_away_goals_per_team[team] = (average_away_goals_per_team[team].to_f / games.size).round(2)
-		end
-		average_away_goals_per_team
-	end
-
-	def home_scores_hash
-		games_grouped_by_home_team = @game_teams_path.group_by {|row| row[:team_id]}
-		average_home_goals_per_team = {}
-
-		games_grouped_by_home_team.each do |team, games|
-			average_home_goals_per_team[team] = 0
-				games.each do |game|
-				average_home_goals_per_team[team] += game[:goals].to_i if game[:hoa] == "home"
-			end
-			average_home_goals_per_team[team] = (average_home_goals_per_team[team].to_f / games.size).round(2)
-		end
-		average_home_goals_per_team
-	end
 
   def highest_scoring_visitor 
 		visitor_highest = visitor_scores_hash.max_by {|k,v| v }
@@ -95,13 +69,6 @@ class GameTeams
     end
   end
 
-  # def games_by_game_id
-  #   #memoization this @games_by_game_id ||= [everything below]
-  #   @games_by_game_id ||= @game_teams_path.group_by do |row| 
-  #     row[:game_id]
-  #   end
-  # end
-
   def game_ids_by_season(season_id)
     games_by_season[season_id].map do |games|
       games[:game_id]
@@ -138,7 +105,6 @@ class GameTeams
 
   def teams_with_tackles(games_array) #HELPER for most and fewest tackles
      hash = Hash.new{|k,v| k[v] = []}
-    # hash = create_hash
     games_array.each do |game_id|
     next if games_by_game_id[game_id].nil?
       games_by_game_id[game_id].each do |game|
