+++ lib/team.rb
@@ -1,25 +1,22 @@
 require 'csv'
+require_relative 'modules/hashable'
 
 class Team
+  include Hashable
   attr_reader :team_path,
               :game_path,
-              :game_teams_path,
-              :team_id,
-              :team_name
+              :game_teams_path
 
   def initialize(file_path, file_path2, file_path3)
     @team_path = file_path
     @game_teams_path = file_path2
     @game_path = file_path3
-    @team_id = file_path[:team_id]
-    @team_name = file_path[:teamname]
   end
 
   def count_of_teams
     @team_path.count
   end
 
-  
   def best_offense
     best = average_goals_by_team_hash.max_by {|k,v| v}
     
@@ -52,35 +49,6 @@ class Team
     end
     average_goals_per_team
   end 
-    
-  
-  def visitor_scores_hash
-    games_grouped_by_away_team = @game_teams_path.group_by {|row| row[:team_id]}
-    average_away_goals_per_team = {}
-
-    games_grouped_by_away_team.each do |team, games|
-      average_away_goals_per_team[team] = 0
-        games.each do |game|
-        average_away_goals_per_team[team] += game[:goals].to_i if game[:hoa] == "away"
-      end
-      average_away_goals_per_team[team] = (average_away_goals_per_team[team].to_f / games.size).round(2)
-    end
-    average_away_goals_per_team
-  end
-
-  def home_scores_hash
-		games_grouped_by_home_team = @game_teams_path.group_by {|row| row[:team_id]}
-		average_home_goals_per_team = {}
-
-		games_grouped_by_home_team.each do |team, games|
-			average_home_goals_per_team[team] = 0
-				games.each do |game|
-				average_home_goals_per_team[team] += game[:goals].to_i if game[:hoa] == "home"
-			end
-			average_home_goals_per_team[team] = (average_home_goals_per_team[team].to_f / games.size).round(2)
-		end
-		average_home_goals_per_team
-	end
 
   def highest_scoring_visitor
 		visitor_highest = visitor_scores_hash.max_by {|k,v| v }
